Quickly completed homework:_)

+++ examples/src/main/java/ru/geekbrains/orm/User.java
@@ -2,23 +2,23 @@ package ru.geekbrains.orm;
 
 public class User {
 
-    private int id;
+    private long id;
 
     private String login;
 
     private String password;
 
-    public User(int id, String login, String password) {
+    public User(long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
